ZEPPELIN-1653: Make UI responsive
When we do the following operations, the entire notebook is transmitted back to client.
1) Insert paragraph
2) Remove paragraph
3) Update paragraph
4) Paragraph status update
5) commit paragraph
When the json becomes larger(which happens when large  output is stored in the json), the time to transfer the notebook to client is directly proportional.(we can check this in chrome browser-> dev console-> select ws request-> on the right pane that opens, select frames tab and observe the length column)

And since the UI update is based on this new transmitted json, the UI seems unresponsive/laggy.
In this PR, Making the updates selective.(Not sending the complete json back to client)

Improvement

NA

ZEPPELIN-1653

-In chrome browser-> dev console-> select ws request-> on the right pane that opens, select frames tab and observe  *NOTE* message.
We should not see the message for the above mentioned operations except when the notebook is loaded is for first time
-    We should not see *GET_NOTE* log message for above operations.

NA

* Does the licenses files need update?No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: karuppayya <karuppayya1990@gmail.com>
Author: Karup <karuppayya@outlook.com>

Closes #1624 from karup1990/ZEPPELIN-1653 and squashes the following commits:

bbdfc9a [karuppayya] Address review comments
4cd9dcd [karuppayya] fix lookandfeel config
c99383a [Karup] Refactor code , dont broadcast note